+++ src/components/DocumentSelector/index.js
@@ -20,12 +20,12 @@ const defaultOptions = (i18n) => [{
 class DocumentSelector extends Component {
 
   getOptions = () => {
-    const {i18n, types} = this.props
+    const {i18n, documentTypes} = this.props
     const defaultDocOptions = defaultOptions(i18n)
-    const options = defaultDocOptions.filter((option) => types.indexOf(option.value) >= 0)
-
-    // If no valid options passed, default to all options
-    return (options.length === 0) ? defaultDocOptions : options
+    
+    const options = defaultDocOptions.filter((option) => documentTypes && documentTypes[option.value])
+    // If no valid options passed, default to defaultDocOptions
+    return options.length ? options : defaultDocOptions
   }
 
   handleSelect = (e, value) => {